# api/openapi.yaml
openapi: 3.0.3
info:
  title: recommend4me
  version: 0.1.0
  description: |
    A local recommender of works — games, books — collected from sites by plugins. The user grades
    some of them; a model learns the order of the grades and puts every other work on the user's
    own 0..10 scale.

    An error is answered as application/problem+json (the Problem schema): its "detail" says, in
    the user's language, what went wrong.
...
paths:
...
  /api/items/{source}/{item}/suggestions:
    get:
      tags: [suggestions]
      operationId: getSuggestions
      summary: The values the work should have, worked out from its text and its other values
      description: |
        For every facet its source suggests values of: the values to add, the likeliest first, and
        the site's values that do not fit. The user confirms one by adding it (correctFacet, added)
        and rejects one by taking it away (correctFacet, not added); an answered value is no longer
        listed. Made at once for a work that never had suggestions.
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
      responses:
        '200':
          description: The suggestions, one entry per facet
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacetSuggestions'
        '404':
          description: No such item
...
  /api/sources/{source}/job:
    post:
      tags: [jobs]
      operationId: startJob
      parameters:
        - $ref: '#/components/parameters/Source'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        '202':
          description: Started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '409':
          description: A job of the source is already running, or the source has no such mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
    delete:
      tags: [jobs]
      operationId: cancelJob
      parameters:
        - $ref: '#/components/parameters/Source'
      responses:
        '200':
          description: The job is being stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
  /api/types/{type}/model:
    get:
      tags: [model]
      operationId: getModel
      parameters:
        - $ref: '#/components/parameters/Type'
      responses:
        '200':
          description: The model of the content type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfo'
  /api/types/{type}/model/train:
    post:
      tags: [model]
      operationId: trainModel
      parameters:
        - $ref: '#/components/parameters/Type'
      responses:
        '200':
          description: The model after training
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfo'
  /api/types/{type}/model/scorer:
    put:
      tags: [model]
      operationId: chooseScorer
      summary: Makes a scorer the one the content type is ranked by, and trains it
      parameters:
        - $ref: '#/components/parameters/Type'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScorerChoice'
      responses:
        '200':
          description: The model after training
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfo'
        '400':
          description: No such scorer
  /api/types/{type}/model/compare:
    post:
      tags: [model]
      operationId: compareScorers
      summary: Cross-validates every scorer on the current grades; nothing is kept
      parameters:
        - $ref: '#/components/parameters/Type'
      responses:
        '200':
          description: The quality of every scorer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScorerQuality'
...
  /api/sources/{source}/settings:
    put:
      tags: [settings]
      operationId: updateSourceSettings
      parameters:
        - $ref: '#/components/parameters/Source'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Setting keys to values; an empty value takes a setting away
              additionalProperties:
                type: string
      responses:
        '200':
          description: The source's settings after the change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceSettings'
...
  /api/pages:
    get:
      tags: [pages]
      operationId: getPage
      summary: What the browser extension shows on a page of a site
      description: |
        How the source marks its pages up, and - on a work's page - the work as the application
        knows it: the prediction, the grade, the facets with the user's corrections, the
        suggestions, the marks of its pictures and reviews.
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageView'
        '404':
          description: No source knows the page
  /api/pages/cards:
    post:
      tags: [pages]
      operationId: getCards
      summary: The prediction and the grade of the works the cards of a list link to
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardsRequest'
      responses:
        '200':
          description: One entry per link the application knows the work of
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardView'
...
components:
  parameters:
    Type:
      name: type
      in: path
      required: true
      schema:
        type: string
    Source:
      name: source
      in: path
      required: true
      schema:
        type: string
    Item:
      name: item
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Grade:
      type: integer
      minimum: 1
      maximum: 5
      description: |
        The user's grade, an order and not a number: 1 - do not like it, 2 - can be played (read),
        3 - liked it on the whole, 4 - very good, 5 - give me more.
...
    SourceMode:
      type: string
      enum: [SCRAPE_ALL, UPDATES, BROWSER]
...
    FacetValueInfo:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
        name:
          type: string
        corrected:
          type: string
          enum: [ADDED, REMOVED]
          description: The user's correction of the site's value; a REMOVED value is shown crossed out
    ItemFacet:
      type: object
      required: [facet, label, values]
      properties:
        facet:
          type: string
          description: The facet's key within its source
        label:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/FacetValueInfo'
...
    RatingRecord:
      type: object
      required: [source, item, version, grade, ratedAt, current]
      properties:
        source:
          type: string
        item:
          type: string
          description: The item of the work graded - this one or one linked to it
        current:
          type: boolean
          description: The grade is of the item's current version
        version:
          type: string
        grade:
          $ref: '#/components/schemas/Grade'
        ratedAt:
          type: string
          format: date-time
    FeatureContribution:
      type: object
      required: [feature, label, contribution, present]
      properties:
        feature:
          type: string
        label:
          type: string
        contribution:
          type: number
          format: double
          description: Points of the 0..10 scale the feature moves the work
        present:
          type: boolean
...
    JobRequest:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: '#/components/schemas/SourceMode'
    JobState:
      type: string
      enum: [IDLE, RUNNING, CANCELLING, DONE, CANCELLED, FAILED]
    JobStatus:
      type: object
      required: [source, state, processed, total, newItems, errors, items]
      properties:
        source:
          type: string
        state:
          $ref: '#/components/schemas/JobState'
        mode:
          $ref: '#/components/schemas/SourceMode'
        resumable:
          $ref: '#/components/schemas/SourceMode'
        phase:
          type: string
        phaseStartedAt:
          type: string
          format: date-time
        processed:
          type: integer
        total:
          type: integer
        newItems:
          type: integer
        errors:
          type: integer
        message:
          type: string
        loggedIn:
          type: boolean
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        items:
          type: integer
          description: Items of the source in the database
        catalogue:
          $ref: '#/components/schemas/CatalogueProgress'
    CatalogueProgress:
      type: object
      required: [stage, done, total, complete]
      description: How far the whole catalogue is loaded, across runs of the job
      properties:
        stage:
          type: string
          description: What the scrape does now or would do next ("feed", "details"); "none" - never started
        done:
          type: integer
        total:
          type: integer
        complete:
          type: boolean
          description: The catalogue is loaded; only updates are left
    ScorerInfo:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string
    ScorerChoice:
      type: object
      required: [scorer]
      properties:
        scorer:
          type: string
    ModelMetrics:
      type: object
      required: [folds, concordance, concordanceSe]
      description: Out-of-fold quality from cross-validation over works
      properties:
        folds:
          type: integer
        concordance:
          type: number
          format: double
          description: Share of pairs of works of different grades put in the right order
        concordanceSe:
          type: number
          format: double
        spearman:
          type: number
          format: double
        auc:
          type: number
          format: double
          description: ROC AUC of liked (above "can be played") against disliked
    ModelCandidate:
      type: object
      required: [parameter, metrics]
      properties:
        parameter:
          type: number
          format: double
        metrics:
          $ref: '#/components/schemas/ModelMetrics'
    ScorerQuality:
      type: object
      required: [scorer, title, candidates]
      properties:
        scorer:
          type: string
        title:
          type: string
        chosen:
          $ref: '#/components/schemas/ModelCandidate'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/ModelCandidate'
    GradeOnScale:
      type: object
      required: [grade, score, ratings]
      properties:
        grade:
          $ref: '#/components/schemas/Grade'
        score:
          type: number
          format: double
        ratings:
          type: integer
    ModelInfo:
      type: object
      required: [trained, training, stale, textEncoderReady, imageEncoderReady, ladder, ratingCounts, features, candidates, topPositive, topNegative, scorers, scorer]
      properties:
        trained:
          type: boolean
        training:
          type: boolean
        stale:
          type: boolean
        textEncoderReady:
          type: boolean
        imageEncoderReady:
          type: boolean
        scorer:
          type: string
          description: The scorer the type is ranked by
        scorers:
          type: array
          items:
            $ref: '#/components/schemas/ScorerInfo'
        trainedAt:
          type: string
          format: date-time
        parameter:
          type: number
          format: double
        ladder:
          type: array
          items:
            $ref: '#/components/schemas/GradeOnScale'
        ratingCounts:
          type: array
          items:
            type: integer
        features:
          type: integer
        metrics:
          $ref: '#/components/schemas/ModelMetrics'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/ModelCandidate'
        topPositive:
          type: array
          items:
            $ref: '#/components/schemas/FeatureContribution'
        topNegative:
          type: array
          items:
            $ref: '#/components/schemas/FeatureContribution'
    SettingValue:
      type: object
      required: [key, label, kind, secret, set]
      properties:
        key:
          type: string
        label:
          type: string
        kind:
          type: string
          enum: [TEXT, NUMBER]
        secret:
          type: boolean
        set:
          type: boolean
        value:
          type: string
          description: The value; never given for a secret
        hint:
          type: string
    SourceSettings:
      type: object
      required: [source, title, values]
      properties:
        source:
          type: string
        title:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/SettingValue'
...
    SuggestedValue:
      type: object
      required: [key, name, chance]
      properties:
        key:
          type: string
        name:
          type: string
        chance:
          type: number
          format: double
          description: The chance the work has the value, 0..1
    FacetSuggestions:
      type: object
      required: [facet, label, suggested, doubted]
      properties:
        facet:
          type: string
          description: The facet's key within the item's source
        label:
          type: string
        suggested:
          type: array
          description: Values to add, the likeliest first
          items:
            $ref: '#/components/schemas/SuggestedValue'
        doubted:
          type: array
          description: The site's values that do not fit the work
          items:
            $ref: '#/components/schemas/SuggestedValue'
    FacetDecorInfo:
      type: object
      required: [facet, values]
      properties:
        facet:
          type: string
        values:
          type: string
    ReviewDecorInfo:
      type: object
      required: [selector, idAttribute, idPrefix]
      properties:
        selector:
          type: string
        idAttribute:
          type: string
        idPrefix:
          type: string
    CardDecorInfo:
      type: object
      required: [selector, link]
      properties:
        selector:
          type: string
        link:
          type: string
    PageDecorInfo:
      type: object
      required: [facets]
      description: CSS selectors of the site's own elements the extension marks up (the plugin API's PageDecor)
      properties:
        panelAfter:
          type: string
        facets:
          type: array
          items:
            $ref: '#/components/schemas/FacetDecorInfo'
        reviews:
          $ref: '#/components/schemas/ReviewDecorInfo'
        pictures:
          type: string
          description: The selector of the work's images on its page
        cards:
          $ref: '#/components/schemas/CardDecorInfo'
    PagePicture:
      type: object
      required: [position, url]
      properties:
        position:
          type: integer
        url:
          type: string
        mark:
          type: integer
    PageItem:
      type: object
      required: [summary, facets, suggestions, explanation, grades, pictures, reviewMarks, ratings]
      properties:
        summary:
          $ref: '#/components/schemas/ItemSummary'
        facets:
          type: array
          description: The facets the page marks up or the application suggests values of, corrections included
          items:
            $ref: '#/components/schemas/ItemFacet'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/FacetSuggestions'
        explanation:
          type: array
          description: What pushed the prediction most, strongest first
          items:
            $ref: '#/components/schemas/FeatureContribution'
        grades:
          type: array
          description: The labels of the grades 1..5 of the work's content type
          items:
            type: string
        pictures:
          type: array
          items:
            $ref: '#/components/schemas/PagePicture'
        reviewMarks:
          type: object
          description: The user's marks of the work's reviews, by review id
          additionalProperties:
            type: integer
        ratings:
          type: array
          items:
            $ref: '#/components/schemas/RatingRecord'
    PageView:
      type: object
      required: [source, sourceTitle, decor]
      properties:
        source:
          type: string
        sourceTitle:
          type: string
        decor:
          $ref: '#/components/schemas/PageDecorInfo'
        itemId:
          type: string
          description: The work the page is of, when it is a work's page - known to the application or not yet
        item:
          $ref: '#/components/schemas/PageItem'
    CardsRequest:
      type: object
      required: [url, links]
      properties:
        url:
          type: string
          description: The address of the page the cards are on
        links:
          type: array
          items:
            type: string
    CardView:
      type: object
      required: [link, source, item, title]
      properties:
        link:
          type: string
          description: The link as it was sent
        source:
          type: string
        item:
          type: string
        title:
          type: string
        grade:
          $ref: '#/components/schemas/Grade'
        prediction:
          type: number
          format: double